// src/interfaces/interface.ts
export type NumberDecimal = {
  value: string
  decimals: number
}
// ...
export type PaginatedRes<T> = {
  result: T[]
  maxItemsCount: number
}

export type PageOptions = {
  limit: number
  skip: number
}